[BUGFIX] Minor code cleanups
A small series of minor cleanups:
* Setting a no longer existing property
* Slight variable type improvements
* Unused variable assignment

Resolves: #87267
Releases: master
Reviewed-on: https://review.typo3.org/59269
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Alexander Schnitzler <typo3@alexanderschnitzler.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>

+++ typo3/sysext/backend/Classes/Controller/EditDocumentController.php
@@ -726,7 +726,7 @@ class EditDocumentController
             || isset($parsedBody['_saveandclosedok'])
         ) {
             // Redirect if element should be closed after save
-            return $this->closeDocument(abs($this->closeDoc), $request);
+            return $this->closeDocument((int)abs($this->closeDoc), $request);
         }
         return null;
     }
